Version 1.7.0

+++ deps/src/libcrypto/bio/b_print.c
@@ -635,7 +635,7 @@ fmtfp(char **sbuffer,
             fvalue = tmpvalue;
     }
     ufvalue = abs_val(fvalue);
-    if (ufvalue > ULONG_MAX) {
+    if (ufvalue > (double)ULONG_MAX) {
         /* Number too big */
         return 0;
     }
